feat(remote): rework brew setup
This reworks the bootstrap to work in the remote configuration setting.
The remote configuration is used to bootstrap a brand new machine. At
that point, there is no assurance the shell is properly setup. Comp
requires the user install brew manually on the machine. This will put
the brew executable in place but not loaded in the shell. The
bootstrapping script and detect this and configure the shell
accordingly.