refactor(shell): split ev & env.d directories
This commit leverages a refactor in dotfiles to support a "env.d"
directory. This way comp can drop in multiple files instead of one.
This allows multiple roles to create shell configs.

The commit also includes a refactoring to extract out variables for the
"env.d" and "ev" directories. It drops the previous "skillshare" ev
directory in favor the common "comp" directory. Technically there is no
need for a separate directory for role because if there is any conflicts
the last applied role wins. In this way the per-host roles are
refinements on the base. The previous "skillshare" ev group was
superfluous.